added validations

+++ db/migrate/20140911112109_create_budgets.rb
@@ -2,7 +2,7 @@ class CreateBudgets < ActiveRecord::Migration
   def change
     create_table :budgets do |t|
       t.string :name
-      t.decimal :b_and_t_rate
+      t.float :b_and_t_rate
       t.timestamps
       t.belongs_to :user
     end
