+++ db/migrate/20140911112118_create_resources.rb
@@ -2,7 +2,7 @@ class CreateResources < ActiveRecord::Migration
   def change
     create_table :resources do |t|
       t.integer :base_salary
-      t.decimal :bonus
+      t.float :bonus
       t.integer :quantity
       t.string :css_left_start
       t.string :css_left_end
